support sending new notes to telegram clients through bot

+++ app/src/main/java/org/flashback/server/handlers/DeleteAccountHandler.java
@@ -12,7 +12,7 @@ public class DeleteAccountHandler {
 
     public static void handle(RequestResponsePair exchange) {
         try{
-            Long userId = Authenticator.authenticate(exchange.getRequest());
+            Integer userId = Authenticator.authenticate(exchange.getRequest());
             Database.deleteUser(userId);
             MessageResponse response = new MessageResponse(true, HttpStatus.NO_CONTENT_204, "deleted");
             GenericHandler.sendResponse(response, exchange);
